Refactor grow
The spec draft's version of grow is kind of hard to understand. Refactor
CAS and load, and spec it as a CAS loop. I *think* the memory model
guarantees remain the same as the current version, even though there are
more events. There's no really other way to implement this anyway AFAIK.