Assert that Integer-Indexed Object isn't out-of-bounds
Because the result of IntegerIndexedObjectLength
is `a non-negative integer of ~out-of-bounds~`,
you need to deal with the `~out-of-bounds~` possibility
before doing any arithmetic on it.

Here, I've simply asserted that the result isn't `~out-of-bounds~`,
but I don't know that that's true.
You might have to do something different (e.g., throw).